Starting with the structure of Django
The first move was to configure the environment. I used the virtual environment to be easier for me to code. Then, I started coding with basic html and css on views file inside of core (which is the application heart.

The main commands that I used were:

Pip install django - To install django
Python -m venv devdjango -> To create the virtual environment
Pip freeze - To check if all the technologies I needed was settled. 
django-admin startproject hellodjango - To start the project named hellodjango
django-admin startapp core - To start the app inside of hellodjango project